+++ envs/realman/rm65b_joystick_env.py
@@ -326,7 +326,7 @@ class RM65BJoystickEnv(OrcaGymRemoteEnv):
                     ee_position,
                     ee_velocity,
                     fingers_width
-                ]).copy()            
+                ], dtype=np.float32).copy()            
         result = {
             "observation": obs,
             "achieved_goal": achieved_goal,
