+++ envs/realman/rm75bv_joystick_env.py
@@ -540,7 +540,7 @@ class RM75BVJoystickEnv(OrcaGymRemoteEnv):
                     ee_position,
                     ee_velocity,
                     fingers_width
-                ]).copy()            
+                ], dtype=np.float32).copy()            
         result = {
             "observation": obs,
             "achieved_goal": achieved_goal,
